Move shared code into a shared module.
There was overlapping/copied code in the es2 and es5 modules.  I have moved them into a shared elasticsearch module to ensure consistency.
