+++ elasticsearch/src/main/java/com/netflix/conductor/elasticsearch/query/parser/FunctionThrowingException.java
@@ -16,25 +16,15 @@
 /**
  * 
  */
-package com.netflix.conductor.dao.es5.index.query.parser;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+package com.netflix.conductor.elasticsearch.query.parser;
 
 /**
  * @author Viren
  *
  */
-public class TestName extends AbstractParserTest {
+@FunctionalInterface
+public interface FunctionThrowingException<T> {
 
-	@Test
-	public void test() throws Exception{
-		String test =  "metadata.en_US.lang		";
-		Name name = new Name(getInputStream(test));
-		String nameVal = name.getName();
-		assertNotNull(nameVal);
-		assertEquals(test.trim(), nameVal);
-	}
+	void accept(T t) throws Exception;
+
 }
