+++ es2-persistence/src/main/java/com/netflix/conductor/dao/es/index/query/parser/GroupedExpression.java
@@ -18,6 +18,9 @@
  */
 package com.netflix.conductor.dao.es.index.query.parser;
 
+import com.netflix.conductor.elasticsearch.query.parser.AbstractNode;
+import com.netflix.conductor.elasticsearch.query.parser.ParserException;
+
 import org.elasticsearch.index.query.QueryBuilder;
 
 import java.io.InputStream;
