+++ es2-persistence/src/main/java/com/netflix/conductor/dao/es/index/query/parser/NameValue.java
@@ -18,6 +18,14 @@
  */
 package com.netflix.conductor.dao.es.index.query.parser;
 
+import com.netflix.conductor.elasticsearch.query.parser.AbstractNode;
+import com.netflix.conductor.elasticsearch.query.parser.ComparisonOp;
+import com.netflix.conductor.elasticsearch.query.parser.ConstValue;
+import com.netflix.conductor.elasticsearch.query.parser.ListConst;
+import com.netflix.conductor.elasticsearch.query.parser.Name;
+import com.netflix.conductor.elasticsearch.query.parser.ParserException;
+import com.netflix.conductor.elasticsearch.query.parser.Range;
+
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
