+++ es5-persistence/build.gradle
@@ -1,13 +1,8 @@
 dependencies {
-
-	compile project(':conductor-core')
-	compile "com.google.inject:guice:${revGuice}"
-
-	//ES5 Dependency
-	compile "org.apache.logging.log4j:log4j-api:${revLog4jApi}"
-	compile "org.apache.logging.log4j:log4j-core:${revLog4jCore}"
+    compile project(':conductor-elasticsearch')
 
 	compile "org.elasticsearch:elasticsearch:${revElasticSearch5}"
 	compile "org.elasticsearch.client:transport:${revElasticSearch5}"
-	compile "commons-io:commons-io:${revCommonsIo}"
+
+    testCompile project(':conductor-elasticsearch').sourceSets.test.output
 }
