+++ es5-persistence/src/main/java/com/netflix/conductor/dao/es5/index/query/parser/Expression.java
@@ -18,12 +18,16 @@
  */
 package com.netflix.conductor.dao.es5.index.query.parser;
 
-import java.io.BufferedInputStream;
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
+import com.netflix.conductor.elasticsearch.query.parser.AbstractNode;
+import com.netflix.conductor.elasticsearch.query.parser.BooleanOp;
+import com.netflix.conductor.elasticsearch.query.parser.ParserException;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
+
+import java.io.BufferedInputStream;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
 
 /**
  * @author Viren
