+++ es5-persistence/src/main/java/com/netflix/conductor/dao/es5/index/query/parser/GroupedExpression.java
@@ -1,67 +1,67 @@
 /**
  * Copyright 2016 Netflix, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 /**
- * 
+ *
  */
 package com.netflix.conductor.dao.es5.index.query.parser;
 
-import java.io.InputStream;
+import com.netflix.conductor.elasticsearch.query.parser.AbstractNode;
+import com.netflix.conductor.elasticsearch.query.parser.ParserException;
 
 import org.elasticsearch.index.query.QueryBuilder;
 
+import java.io.InputStream;
+
 /**
  * @author Viren
- * 
+ *
  */
 public class GroupedExpression extends AbstractNode implements FilterProvider {
 
-	private Expression expression;
-	
-	public GroupedExpression(InputStream is) throws ParserException {
-		super(is);
-	}
+    private Expression expression;
 
-	@Override
-	protected void _parse() throws Exception {
-		byte[] peeked = read(1);
-		assertExpected(peeked, "(");
-		
-		this.expression = new Expression(is);
-		
-		peeked = read(1);
-		assertExpected(peeked, ")");
-		
-	}
-	
-	@Override
-	public String toString(){
-		return "(" + expression + ")";
-	}
+    public GroupedExpression(InputStream is) throws ParserException {
+        super(is);
+    }
 
-	/**
-	 * @return the expression
-	 */
-	public Expression getExpression() {
-		return expression;
-	}
+    @Override
+    protected void _parse() throws Exception {
+        byte[] peeked = read(1);
+        assertExpected(peeked, "(");
 
-	@Override
-	public QueryBuilder getFilterBuilder() {
-		return expression.getFilterBuilder();
-	}
+        this.expression = new Expression(is);
 
-	
+        peeked = read(1);
+        assertExpected(peeked, ")");
+
+    }
+
+    @Override
+    public String toString() {
+        return "(" + expression + ")";
+    }
+
+    /**
+     * @return the expression
+     */
+    public Expression getExpression() {
+        return expression;
+    }
+
+    @Override
+    public QueryBuilder getFilterBuilder() {
+        return expression.getFilterBuilder();
+    }
+
+
 }
