+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name='conductor'
 
-include 'client','common','contribs','core','es2-persistence','es5-persistence','jersey','mysql-persistence'
-include 'redis-persistence','server','test-harness','ui'
+include 'client','common','contribs','core', 'elasticsearch', 'es2-persistence','es5-persistence','jersey'
+include 'mysql-persistence', 'redis-persistence','server','test-harness','ui'
 include 'protogen'
 include 'grpc', 'grpc-server', 'grpc-client'
